+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 - The `config.license.link` parameter has been changed to `config.licenseLink`.
+- The required Node.js version is now 18.18.0 or higher.
 
 ### Fixed
 - Corrected a typo in this changelog.