+++ CHANGELOG.md
@@ -5,8 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.1.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-[Unreleased]
-------------
+[1.0.0] - 2024-10-03
+--------------------
 ### Added
 - More usage examples to the readme.
 
@@ -142,7 +142,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Initial release.
 
-[Unreleased]: https://github.com/jbenner-radham/node-readme-md/compare/v0.15.0...HEAD
+[Unreleased]: https://github.com/jbenner-radham/node-readme-md/compare/v1.0.0...HEAD
+[1.0.0]: https://github.com/jbenner-radham/node-readme-md/compare/v0.15.0...v1.0.0
 [0.15.0]: https://github.com/jbenner-radham/node-readme-md/compare/v0.14.0...v0.15.0
 [0.14.0]: https://github.com/jbenner-radham/node-readme-md/compare/v0.13.1...v0.14.0
 [0.13.1]: https://github.com/jbenner-radham/node-readme-md/compare/v0.13.0...v0.13.1